bugfix: wrong number type in AlgoChiSquaredTest
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@21716 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/kernel/statistics/AlgoChiSquaredTest.java
@@ -190,12 +190,12 @@ public class AlgoChiSquaredTest extends AlgoElement {
 			// compute expected values if list2 is not given
 			if (geoList2 == null) {
 
-				int[] columnSum = new int[columns];
+				double[] columnSum = new double[columns];
 				for (int j = 0; j < columns; j++) {
 					columnSum[j] = 0;
 				}
 
-				int[] rowSum = new int[rows];
+				double[] rowSum = new double[rows];
 				for (int i = 0; i < rows; i++) {
 					rowSum[i] = 0;
 				}
